Clean up styling, data types, add RcppParallel

+++ src/mcmc_utils.h
@@ -389,7 +389,7 @@ inline float logSumExp(const std::vector<float> &x)
 
     float sum = 0;
 #pragma omp simd reduction(+ : sum)
-    for (int i = 0; i < x.size(); ++i)
+    for (size_t i = 0; i < x.size(); ++i)
     {
         sum += std::exp(x[i] - max_el);
     }
